Removed dependency on cflib for promise / tailcall.  Converted to plain continuation passing style and a trampoline that checks for typeof res === "function" instead of instanceof TailCall.  (Mostly for performance reasons).
Changed implementation of async functions from those returning a Promise to
those registered with a different function (registerAsyncCommand), which gets
a continuation function passed as the first argument (and must not return a
value).

Removed support for thisobj in registered functions (fn.apply is much slower
than fn()).

Various optimizations and improvements to error handling.